feat: add loading and error toaster

+++ features/content/content_common/src/main/java/com/tokopedia/content/common/producttag/data/ProductTagRepositoryImpl.kt
@@ -132,7 +132,7 @@ class ProductTagRepositoryImpl @Inject constructor(
         }
     }
 
-    override suspend fun getShopInfoByID(shopIds: List<Int>): ShopUiModel {
+    override suspend fun getShopInfoByID(shopIds: List<Long>): ShopUiModel {
         return withContext(dispatchers.io) {
             val response = getShopInfoByIDUseCase.apply {
                 setRequestParams(GetShopInfoByIDUseCase.createParams(
